Changed file import to use columns [#965]
