+++ comixed-webui/src/app/comic-files/components/comic-file-covers/comic-file-covers.component.scss
