--- comixed-webui/src/app/comic-files/components/comic-file-list/comic-file-list.component.html
@@ -1,45 +1,0 @@
-<mat-table
-  aria-label="Comic files available for import"
-  matSort
-  [dataSource]="dataSource"
-  (matSortChange)="onSortChanged($event)"
->
-  <ng-container matColumnDef="selected">
-    <mat-header-cell mat-sort-header *matHeaderCellDef>
-      <mat-checkbox
-        [checked]="allSelected"
-        [disabled]="this.files.length === 0"
-        (change)="onSelectAll($event.checked)"
-      ></mat-checkbox>
-    </mat-header-cell>
-    <mat-cell *matCellDef="let file">
-      <mat-checkbox
-        [checked]="selectedFiles.includes(file)"
-        (click)="$event.stopPropagation()"
-        (change)="onSelectFile(file, $event.checked)"
-      ></mat-checkbox>
-    </mat-cell>
-  </ng-container>
-
-  <ng-container matColumnDef="filename">
-    <mat-header-cell mat-sort-header *matHeaderCellDef>
-      {{ "comic-files.header.filename" | translate }}
-    </mat-header-cell>
-    <mat-cell *matCellDef="let file">{{ file.baseFilename }}</mat-cell>
-  </ng-container>
-
-  <ng-container matColumnDef="size">
-    <mat-header-cell mat-sort-header *matHeaderCellDef>
-      {{ "comic-files.header.size" | translate }}
-    </mat-header-cell>
-    <mat-cell *matCellDef="let file">
-      {{ file.size / 1024 | number: "1.2-2" }}
-    </mat-cell>
-  </ng-container>
-
-  <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-  <mat-row
-    *matRowDef="let row; columns: displayedColumns"
-    (click)="onRowSelected(row)"
-  ></mat-row>
-</mat-table>
