+++ comixed-webui/src/app/comic-files/pages/import-comics-page/import-comics-page.component.html
@@ -7,8 +7,6 @@
     [user]="user"
     [comicFiles]="files"
     [selectedComicFiles]="selectedFiles"
-    [ignoreMetadata]="ignoreMetadata"
-    [deleteBlockedPages]="deleteBlockedPages"
   ></cx-comic-file-toolbar>
   <h2>
     {{
@@ -16,10 +14,9 @@
         | translate: { count: files.length, selected: selectedFiles.length }
     }}
   </h2>
-  <cx-comic-file-list
+  <cx-comic-file-covers
     *ngIf="files.length > 0"
     [files]="files"
     [selectedFiles]="selectedFiles"
-    (currentFile)="onCurrentFile($event)"
-  ></cx-comic-file-list>
+  ></cx-comic-file-covers>
 </div>
